text-to-speech-bot/requirements.txt
python-telegram-bot==20.3
gTTS==2.3.1
python-dotenv==1.0.0
